+++ libexec/rtld-elf/amd64/reloc.c
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
- * $FreeBSD: src/libexec/rtld-elf/amd64/reloc.c,v 1.15 2004/08/03 08:50:59 dfr Exp $
+ * $FreeBSD: src/libexec/rtld-elf/amd64/reloc.c,v 1.15.8.1 2005/12/30 22:13:56 marcel Exp $
  */
 
 /*
@@ -312,9 +312,9 @@ reloc_non_plt(Obj_Entry *obj, Obj_Entry *obj_rtld)
 	/* missing: R_X86_64_GOTPCREL, R_X86_64_32, R_X86_64_32S, R_X86_64_16, R_X86_64_PC16, R_X86_64_8, R_X86_64_PC8 */
 
 	    default:
-		_rtld_error("%s: Unsupported relocation type %d"
+		_rtld_error("%s: Unsupported relocation type %u"
 		  " in non-PLT relocations\n", obj->path,
-		  ELF_R_TYPE(rela->r_info));
+		  (unsigned int)ELF_R_TYPE(rela->r_info));
 		goto done;
 	    }
 	}
